Add debug bit vmod_so_keep This makes it possible to load VMOD shared objects into a debugger after a varnishd crash.

+++ bin/varnishd/mgt/mgt_vcl.c
@@ -272,7 +272,8 @@ mgt_vcl_del(struct vclprog *vp)
 		FREE_OBJ(vd);
 
 		if (VTAILQ_EMPTY(&vf->vcls)) {
-			AZ(unlink(vf->fname));
+			if (! MGT_DO_DEBUG(DBG_VMOD_SO_KEEP))
+				AZ(unlink(vf->fname));
 			VTAILQ_REMOVE(&vmodhead, vf, list);
 			free(vf->fname);
 			FREE_OBJ(vf);
